:wrench: Update: Switched to npm
