+++ package.json
@@ -7,7 +7,7 @@
     "start": "next start",
     "lint": "next lint",
     "buildcss": "npx tailwindcss -o build.css --minify",
-    "ci": "yarn install --immutable --immutable-cache --check-cache"
+    "ci": "npm install --immutable --immutable-cache --check-cache"
   },
   "dependencies": {
     "framer-motion": "6.2.1",
@@ -33,6 +33,6 @@
     "typescript": "4.5.5"
   },
   "volta": {
-    "node": "16.14.0"
+    "node": "14.19.0"
   }
 }
