+++ pages/index.server.tsx
@@ -1,4 +1,3 @@
-import { StrictMode } from 'react'
 import type { NextPage } from 'next'
 import Parallax from '../components/Parallax.client'
 import Page from '../components/Page.server'
@@ -6,7 +5,6 @@ import Content from '../components/Content.server'
 
 const Home: NextPage = () => {
   return (
-    <StrictMode>
     <div className="h-[4000px]">
       <Page title="Welcome to my Portfolio" header={<Parallax />}>
         <Content>
@@ -15,7 +13,6 @@ const Home: NextPage = () => {
         <div className="h-[400px] w-full bg-red-400" />
       </Page>
     </div>
-    </StrictMode>
   )
 }
